Refactor demos into dedicated examples
